+++ wxc_sdk/as_api.py
@@ -14319,13 +14319,15 @@ class AsWebexSimpleApi:
     #: :class:`AsRestSession` used for all API requests
     session: AsRestSession
 
-    def __init__(self, *, tokens: Union[str, Tokens] = None, concurrent_requests: int = 10):
+    def __init__(self, *, tokens: Union[str, Tokens] = None, concurrent_requests: int = 10, retry_429: bool = True):
         """
 
         :param tokens: token to be used by the API. Can be a :class:`tokens.Tokens` instance, a string or None. If
             None then an access token is expected in the WEBEX_ACCESS_TOKEN environment variable.
         :param concurrent_requests: number of concurrent requests when using multi-threading
         :type concurrent_requests: int
+        :param retry_429: automatically retry for 429 throttling response
+        :type retry_429: bool
         """
         if isinstance(tokens, str):
             tokens = Tokens(access_token=tokens)
@@ -14336,7 +14338,7 @@ class AsWebexSimpleApi:
                                  'WEBEX_ACCESS_TOKEN environment variable')
             tokens = Tokens(access_token=tokens)
 
-        session = AsRestSession(tokens=tokens, concurrent_requests=concurrent_requests)
+        session = AsRestSession(tokens=tokens, concurrent_requests=concurrent_requests, retry_429=retry_429)
         self.attachment_actions = AsAttachmentActionsApi(session=session)
         self.cdr = AsDetailedCDRApi(session=session)
         self.devices = AsDevicesApi(session=session)