+++ docs/changes.rst
@@ -16,6 +16,7 @@ Release history
 - fix: camelCase issues for timezone when updating a location (temp fix): :meth:`wxc_sdk.locations.LocationsApi.update`
 - feat: new example catch_tns.py
 - feat: better handling of CDRs in :class:`wxc_sdk.cdr.CDR` to allow deserialization of addtl. fields
+- feat: new parameter ´retry_429' for :class:`wxc_sdk.WebexSimpleApi and :class:`wxc_sdk.as_api.AsWebexSimpleApi`
 
 1.14.1
 ------